+++ app/Http/Controllers/HelpController.php
@@ -3,17 +3,15 @@
 namespace App\Http\Controllers;
 
 use Illuminate\View\View;
-use App\Models\FAQ;
+use App\Models\Faq;
 
-class HelpController extends Controller
-{
+class HelpController extends Controller {
     /**
      * Display the help view.
      */
-    public function index(): View
-    {
+    public function index(): View {
         return view('help', [
-            'FAQList' => FAQ::$FAQList,
+            'FaqList' => Faq::all(),
         ]);
     }
 }