+++ app/Http/Requests/ProfileUpdateRequest.php
@@ -6,19 +6,22 @@ use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 
-class ProfileUpdateRequest extends FormRequest
-{
+class ProfileUpdateRequest extends FormRequest {
     /**
      * Get the validation rules that apply to the request.
      *
      * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
      */
-    public function rules(): array
-    {
+    public function rules(): array {
         return [
-            'name' => ['required', 'string', 'max:255'],
+            'display_name' => ['nullable', 'string', 'max:255'],
+            'name' => ['nullable', 'string', 'max:255'],
+            'avatar' => ['nullable', 'image', 'max:2048'],
+            'role' => ['nullable', 'numeric', 'between:1,2'],
+            'birthday' => ['nullable', 'string', 'max:255'],
+            'about_me' => ['nullable', 'string', 'max:1080'],
             'email' => [
-                'required',
+                'nullable',
                 'string',
                 'lowercase',
                 'email',