+++ app/Models/User.php
@@ -7,8 +7,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 
-class User extends Authenticatable
-{
+class User extends Authenticatable {
     /** @use HasFactory<\Database\Factories\UserFactory> */
     use HasFactory, Notifiable;
 
@@ -18,8 +17,13 @@ class User extends Authenticatable
      * @var list<string>
      */
     protected $fillable = [
+        'display_name',
         'name',
         'email',
+        'avatar',
+        'role',
+        'birthday',
+        'about_me',
         'password',
     ];
 
@@ -38,8 +42,7 @@ class User extends Authenticatable
      *
      * @return array<string, string>
      */
-    protected function casts(): array
-    {
+    protected function casts(): array {
         return [
             'email_verified_at' => 'datetime',
             'password' => 'hashed',