+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -4,17 +4,22 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
+    public function up(): void {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->string('name');
+            $table->string('display_name')->default('Joe Shmoe');
+            $table->string('name')->unique();
             $table->string('email')->unique();
+            $table->string('avatar')->nullable();
+            $table->string('birthday')->nullable();
+            $table->string('about_me')->nullable();
+            $table->integer('role')->default(2);
+            // 1 = Administrator
+            // 2 = Regular user
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
@@ -40,8 +45,7 @@ return new class extends Migration
     /**
      * Reverse the migrations.
      */
-    public function down(): void
-    {
+    public function down(): void {
         Schema::dropIfExists('users');
         Schema::dropIfExists('password_reset_tokens');
         Schema::dropIfExists('sessions');