+++ resources/views/auth/login.blade.php
@@ -10,7 +10,7 @@
 			<div class="grid-item-1">
 				<x-input-label for="email" :value="__('Email')" />
 				<x-text-input id="email" type="email" name="email" :value="old('email')" required autofocus
-					autocomplete="username" />
+					autocomplete="email" />
 				<x-input-error :messages="$errors->get('email')" />
 			</div>