+++ resources/views/auth/verify-email.blade.php
@@ -1,7 +1,7 @@
 <x-app-layout>
-	<div>
+	<p>
 		{{ __('Thanks for signing up! Before getting started, could you verify your email address by clicking on the link we just emailed to you? If you didn\'t receive the email, we will gladly send you another.') }}
-	</div>
+	</p>
 
 	@if (session('status') == 'verification-link-sent')
 		<div>