+++ resources/views/components/input-label.blade.php
@@ -1,5 +1,8 @@
-@props(['value'])
+@props(['value', 'isRequired' => 'false'])
 
-<label>
+<label class="{{ $attributes['class'] }}" for="{{ $attributes['for'] }}">
 	{{ $value ?? $slot }}
+	@if ($isRequired === true)
+		*
+	@endif
 </label>