+++ resources/views/discussion-board.blade.php
@@ -1,3 +1,7 @@
+@php
+	use App\Models\User;
+@endphp
+
 <x-app-layout>
 	<x-slot name="header">
 		<h2>{{ __('Discussion') }}</h2>
@@ -5,6 +9,12 @@
 
 	<section>
 		<p>Hello From Discussion Board!</p>
+		@php
+			$users = User::all();
+		@endphp
+		@foreach ($users as $user)
+			<p>{{ $user->display_name }}</p>
+		@endforeach
 	</section>
 
 	<x-slot name="footer"></x-slot>