+++ resources/views/profile/edit.blade.php
@@ -3,6 +3,8 @@
 		<h2>
 			{{ __('Profile') }}
 		</h2>
+		<img class="avatar avatar-medium" src="{{ Storage::url(Auth::user()->avatar ?? '/avatars/default-avatar.png') }}"
+			alt="Avatar">
 	</x-slot>
 
 	@include('profile.partials.update-profile-information-form')